requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pyyaml==6.0.1
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1

# AI/ML
openai==1.3.7
anthropic==0.7.8

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Async support
asyncio-mqtt==0.16.1

# Logging and monitoring
structlog==23.2.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Docker and deployment
gunicorn==21.2.0

# Future integrations (for later phases)
# slack-sdk==3.26.1
# PyGithub==1.59.1 